mostly builds except for server
