+++ package.json
@@ -105,6 +105,7 @@
   },
   "dependencies": {
     "@microsoft/powerquery-parser": "0.1.40",
-    "vscode-languageclient": "6.1.3"
+    "vscode-languageclient": "6.1.3",
+    "vscode-languageserver-textdocument": "1.0.1"
   }
 }
